refactor: extract read-query API into collection/query.rs (#577 item 17)
Third sub-step of splitting collection.rs's 5 bundled domains (#577):
the read-query API (search, browse-by-album/artist/genre/decade/BPM,
home-screen sections, artist-profile reads) moves into a
collection::query submodule as a second `impl CollectionScanner`
block, alongside its private helper functions
(group_songs_into_home_items, attach_album_ratings,
get_artist_profile_conn/set_artist_profile_conn/
get_all_artist_profiles_conn, group_by_play_context,
home_item_for_context, get_playlists_by_ids, get_songs_by_ids,
multi_value_contains_sql/pattern, home_item_select_cols) and their
tests. collection.rs keeps scanning/directory management, tag
reading, DB upsert, and the SQL column helpers/mode_query_fragments
that stay crate-visible for other modules (tags.rs, player.rs,
playlist.rs, commands/*.rs).

No behavior change — pure move. Multiple impl blocks for the same
struct across files is idiomatic Rust; CollectionScanner's private
`db` field stays accessible from query.rs since it's a descendant
module of collection.
